First commit
This is the initial upload of this editor. It is still a baby... Not even. It still hasn't been conceived yet.

This is its very bare bones state. It works. You can make maps, you can load maps(as long as the name space is 'map.json', and its in a specific folder relative to the path of the file.)