Rusty Level Editor V0.02
I added a very messy UI for selecting the tile that you want to use. Scrolling no longer works in this version. I will be replacing to functionality to being able to zoom in/out on the canvas. I might re-implement tile scrolling in the future with a hotkey if i feel myself missing it